+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.6.4 (WIP)
 
 * Add more wrappers to the content of a `Link` widget
+* Use the new `alchemy.getRoutes()` method
 
 ## 0.6.3 (2023-02-26)